requirements.txt
Flask
openai
requests
python-dotenv
google-cloud-storage
